`StoreKitObserverModeIntegrationTests`: fixed and disable SK2 `testPurchaseInDevicePostsReceipt`
Turns out this test was only passing because it was waiting for the receipt to be posted during a renewal, not the initial purchase!
This updates the test to reflect the correct expectation (disabling renewals), and added the `XCTExpectFailure` so we can detect if it starts passing in a future iOS version.