Improved `finishAllUnfinishedTransactions`
Also hoping that this would work around `FB11799675`, but it still didn't. I still think it's useful to ensure a clean slate in tests.

This new implementation does 3 things:
- Delete transactions from `SKTestSession.allTransactions()` (those should be empty after `SKTestSession.clearTransactions()`, but this method further ensures that)
- Finish transactions from `StoreKit.Transaction.unfinished`
- Delete transactions in `StoreKit.Transaction.unfinished`

Because this new implementation requires an `SKTestSession`, it exposed the fact that we were calling this from `BaseBackendIntegrationTests`, which not all subclasses create an `SKTestSession`. Now it's in `BaseStoreKitIntegrationTests` where it belongs.