Removed `preventPurchasePopupCallFromTriggeringCacheRefresh`, update caches on `willEnterForeground`
This method was a hack to deal with apps getting `didBecomeActive` notifications in the middle of a purchase when the popups were displayed.
This had several problems:
- It was only being used for StoreKit 1 purchases
- It wasn't used for StoreKit 2 purchases through the SDK
- It wasn't used for purchases initiated outside the SDK (like using the upcoming paywall screens)
- It updated the cache timestamp to "now", including offerings, despite not actually updating data
- It was an indirect and implicit way of solving a problem instead of an explicit one

This solves the same by observing `willEnterForeground`, which only happens when the app actually comes back from the background, and not when it simply loses focus.