`CI`: removed `carthage_archive` from `release` lane
It's useful to have it in `release_checks` lane, to make sure that our framework can be built with `Carthage`.
However, what we actually release is using `export_xcframework`, which also builds the frameworks.

This will make `release` MUCH faster since we no longer build everything twice as many times (and for each platform / architecture combination).