`TestLogHandler`: changed all tests to explicitly uninitialize it
We get a lot of flaky failures in CI due to `TestLogHandler` exceeding its duration. I haven't been able reproduce that locally, but it's likely due to the fact that it relies on the autorelease pool releasing the instance, which in turn removes the `LogMessageObserver`.
To fix that, this PR avoids creating custom `TestLogHandler` that rely on the lifetime of the test, and instead explicitly set it to `nil` in `tearDown`.

I implemented that in `TestCase` (now also available for backend tests) so all our test classes can take advantage of that.