Replaced `FatalErrorUtil` with `Nimble`
Turns out that all the testing issues in CI with #2780 were due to these functions.

/Users/ignacio/Desktop/Screenshot 2023-07-12 at 16.45.37.png
Even though I wasn't able to reproduce the slowness, I was able to detect the root cause using Instruments, which detects this "severe hang".
With this commit, all the slow test issues go away :tada:

The only downside is we lose the ability to verify the assertion message.
The benefit is we don't have to maintain that hacky code and instead rely on `Nimble` with a much better implementation that use this hack:
```swift
func unreachable() {
    repeat {
        NSRunLoop.currentRunLoop().run()
    } while (true)
}
```