`Paywalls`: added support for purchasing
This adds a new `PurchaseHandler` type, which can be instantiated with `Purchases` or with a mock implementation for previews.
That gets injected, and templates can now purchase and dismiss themselves.

I've improved the `SimpleApp` setup to better work with the lifetime of purchases. Now it only displays the paywall if the user doesn't have an entitlement.