[move] `Purchases.checkTrialOrIntroDiscountEligibility(packages: [Package])`
This new overload helps the implementation of `RevenueCatUI`, so we can get a complete dictionary of all requested `Packages`, regardless of whether they might refer to the same products, or if the implementation does not contain that in the result.

I also implemented `IntroEligibility`'s `Hashable`'s overrides as well as `debugDescription`, which were needed to debug this.
Turns out there was no test coverage for the other check trial eligibility methods, so this tests those indirectly as well.