`Paywalls`: fixed `{{ total_price_and_per_month }}
We were relying on both the price and price per month to be formatted the same way. When debugging in the simulator, the actual region doesn't change, so StoreKit's internal price formatter doesn't pick it up.
To make this more resilient, this now checks whether it's a monthly subscription instead of comparing strings.