`CachingProductsManager`: removed duplicate log and added tests (#2898)
I noticed this was being logged twice when requesting SK1 products,
which makes it seem like we're fetching them twice:
```
[store_kit] DEBUG: ℹ️ No existing products cached, starting store products request for: ["com.revenuecat.monthly_4.99.1_week_intro"]
```

`CachingProductsManager` was introduced in #1907. Explained there is why
`ProductsFetcherSK1` keeps some caching behavior.
But the main caching layer is now `CachingProductsManager`, so it makes
sense to keep the log only there.

This adds 4 basic integration tests. We had independent coverage of
`ProductsManager` and `CachingProductsManager` with a mock
`ProductsManager`, but no unit tests covering both together. This serves
that purpose too.