`Paywalls`: remove duplicate `RevenueCat` scheme to fix Carthage
`release-checks` was broken because `carthage build --no-skip-current` was failing to find a scheme.
See https://circleci.com/gh/RevenueCat/purchases-ios/113398

After looking into `Carthage's` source code, I saw that it uses `xcodebuild -list`. This was the output:

```
$ xcodebuild -list
 Schemes:
        BackendIntegrationTests
        ObjCAPITester
        ObjCAPITester
        ReceiptParser (purchases-ios)
        ReceiptParser (RevenueCat project)
        ReceiptParserAPITester
        ReceiptParserAPITester
        RevenueCat (purchases-ios)
        RevenueCat (RevenueCat project)
        RevenueCat_CustomEntitlementComputation
        RevenueCatUI
        SwiftAPITester
        SwiftAPITester
        UnitTestsHostApp
        v3LoadShedderIntegration
```

The double `RevenueCat` scheme comes from this, which was introduced in the `paywalls` branch.

We don't need to save this in Git since we have the project's scheme already.