`Xcode`: removed `purchases-ios` SPM reference
The project no longer depends on `purchases-ios` through SPM.
This was needed for 2 reasons:
- `v3LoadShedderIntegration`
- `BackendCustomEntitlementsIntegrationTests`

Both of these depend on `RevenueCat_CustomEntitlementComputation`, which is only an SPM library.

This solution allows us to build it but without having to clone the repo separately through a dependency, and instead embedding a reference to the _local_ Package.
`v3LoadShedderIntegration` depended on the 3.x release, so I've simply removed it from the workspace. It can still be modified and ran directly by opening the project.