`Integration Tests`: fixed more flaky failures (#3218)
Same as #3167. These can fail when verifying that no transactions are
finished because there might be leftover transactions before the test
even begins. In order to ignore those, we clear `TestLogHandler` at the
beginning. We don't do that for every test because in some test we want
to look at what was logged during SDK initialization.