`Paywalls`: simplified `LoadingPaywallView`
This is a better fix than #3235. Instead of marking the view as disabled, which is also used by `PurchaseHandler` to signal that a purchase is in progress, we simply make that view not accept touches.
Which has the actual effect that we want (not being interactive), without messing with how it's displayed).