`StoreKit 1`: disabled `finishTransactions` log on observer mode
Follow up to #3300.
Thanks to @aboedo for pointing out that this should never be logged during observer mode because we'll never have completion blocks in that case.

I still left the "no callbacks to notify" part because it's useful to verify that the SDK is ignoring the transaction notification, but removed the warning that won't be relevant in this case.