`PaywallEventStore`: also remove legacy `revenuecat` documents directory (#3317)
Fixes #3316. Follow up to #3289.

That previous fix was only removing the store itself, but not the
`revenuecat` folder.

Running an app with 4.27.2 produces this:
```bash
$ tree -L 3 ~/Library/Developer/CoreSimulator/Devices/.../data/Containers/Data/Application/.../
~/Library/Developer/CoreSimulator/Devices/.../data/Containers/Data/Application/.../
├── Documents
│   └── revenuecat
│       └── paywall_event_store
```

After this fix, the Documents directory is empty.
```bash
tree -L 4 ~/Library/Developer/CoreSimulator/Devices/.../data/Containers/Data/Application/.../
/Users/ignacio/Library/Developer/CoreSimulator/Devices/.../data/Containers/Data/Application/.../
├── Documents
├── Library
│   ├── Application Support
│   │   └── revenuecat
│   │       └── paywall_event_store
```

This is still covered by the same test introduced in #3289.