`ProductFetcherSK1`: enable `TimingUtil` log
`ProductFetcherSK1` had 2 public methods for fetching products: `products(withIdentifiers:completion:)` and `sk1Products(withIdentifiers:completion:)`.
Only one of them was using `TimingUtil` to log slow requests, and that one happened to not be used by `ProductsManager`, which means that we weren't getting these logs.

This is evident in https://github.com/RevenueCat/react-native-purchases/issues/755. With this change, we'll now be able to see how long the product request took.