`Paywalls`: always dismiss paywalls automatically after a purchase
Fixes #3516.

This logic only existed for `.presentPaywallIfNeeded`. However, it should apply to `PaywallView` when presented manually as well.
This is also consistent with how Android works, reduces boilerplate code required by users, and removes the possibility that a paywall is still presented after a purchase, which doesn't make sense since our paywalls don't have a mode for "already subscribed".