`Paywalls`: remove unscrollable spacing in Template 5 (#3562)
By @thillsman. See #3561. 

### Motivation
As I was implementing this template (with longer content that needs
scrolling), the spacing cut off the headline in a weird way. It also
takes up a small bit of vertical space unnecessarily.

### Description
Instead of doing the previous layout: `header image / even space /
scrollable content / even space / footer`, I updated the layout to make
the header image and scrollable content in a 0 spacing VStack, with the
same existing padding _inside_ the scrollable content to have the same
initial layout as before. Probably best described with two videos:

#### Before:


https://github.com/RevenueCat/purchases-ios/assets/997637/0961e99c-8a25-4099-a763-6838ff1ac0ed

#### After:


https://github.com/RevenueCat/purchases-ios/assets/997637/22274ed2-3ab1-4fc1-894b-83cd8013bce8

---------

Co-authored-by: Tyler Hillsman <tyler.hillsman@gmail.com>