`Paywalls`: new `presentationMode` parameter (#3632)
### TLDR
- Added a way to present `PaywallView` either as sheet or full screen
- No breaking changes have been made

### Motivation
I wanted to easily present the PaywallView on full screen.

### Description
I added a new enum `PaywallPresentationMode` which currently has 2
options: `.sheet` and `.fullScreen `. I added this as a parameter to the
current public `presentPaywallIfNeeded` modifiers. The default is
`sheet` so there are no breaking changes with the current public API.

| Sheet             | Full Screen    |
| ------------- | ------------- |
| ![Simulator Screenshot - iPhone 15 Pro - 2024-02-05 at 18 08
51](https://github.com/RevenueCat/purchases-ios/assets/1515520/80f4c418-5dd7-4ba9-87a3-1b824bd6fc10)
| ![Simulator Screenshot - iPhone 15 Pro - 2024-02-05 at 18 44
15](https://github.com/RevenueCat/purchases-ios/assets/1515520/ab05b52b-6132-41d0-a48b-cecf5b857eeb)
|

PD: I looked for info on how to contribute but it's unclear if I have to
open an issue first or not (contributing says open issue, then when
opening an issue it says please don't open but contact us at
https://app.revenuecat.com/settings/support). I hope this is valid,
otherwise let me know how I can proceed.

---------

Co-authored-by: Luis Ascorbe <luis@caribepixels.com>
Co-authored-by: NachoSoto <NachoSoto@users.noreply.github.com>