`Paywalls`: prioritize `Locale.current` over `Locale.preferredLocales`
Fixes #3655.
Follow up to #3633.

This improves upon that PR by using `Locale.current.removingRegion` **before** trying to look up `Locale.preferredLocales`.

This fixes this scenario:
- Configure your phone with: `en_UK` and `es_ES`
- Launch paywall with localization `en_US` and `es_ES`

Prior to this change, we'd be looking up localizations in this order:
- `en_UK`
- `es_ES`
- `en` (no region)

Because there _is_ a localization for `es_ES` we'd return that one before attempting to look up `Locale.current.removingRegion`.