Remove `restoreStarted` from extensions (#3698)
Since we are deprecating the old function in
https://github.com/RevenueCat/purchases-ios/pull/3693/files

I thought it will be simpler to just add both `restoreStarted` and the
new `purchaseStarted(package:)` in the same release. We haven't made a
release since I merged #3694 so these are safe to remove for now.

I open this PR in case we want to make a release before #3693 is
finished.