[EXTERNAL] Fix Typos in ReceiptStrings.swift (#3756)
<!-- Thank you for contributing to Purchases! Before pressing the
"Create Pull Request" button, please provide the following: -->

### Checklist
- [x] If applicable, unit tests
- [ ] If applicable, create follow-up issues for `purchases-android` and
hybrids

### Motivation
I wanted to fix simple typos in log messages to provide a slightly
better developer experience.

### Description
The changes include a fix for a typo in a string and a typo in an
enumeration case declaration. Unit tests were not implemented for
`ReceiptStrings.swift`, so I added unit tests to cover all
`ReceiptStrings` members.

### Notes
I looked for similar issues in the repositories for `purchases-android`
and hybrids, but I did not find any.