Deprecate observer mode in 5.0 branch (#3956)
Adds a `PurchasesAreCompletedBy` enum which is used to set if the RC SDK
performs the purchase, formerly referred to as "observer mode". This
replaces observer mode and finishTransactions. The change has only been
made to public-facing APIs, internal ones will be renamed as part of a
subsequent PR.

Other changes:

- Renamed `handleObserverModeTransaction(_ PurchaseResult:` to
`recordPurchase(_ PurchaseResult:` to try to better communicate what the
purpose of the call is. Feedback on if this is needed and what the new
name should be is welcome!

- There is a "deprecated" method that doesn't appear to really be a
deprecated method, but the deprecation is used as a way of generating a
warning where customers are likely mis-using the API. I updated its
signature to match the new API. Screenshot of the change for clarity:
<img width="906" alt="image"
src="https://github.com/RevenueCat/purchases-ios/assets/109382862/a7780f50-7f85-443f-8312-b0ef4a0b1928">

---------

Co-authored-by: Will Taylor <wtaylor151@gmail.com>