Reduce CI jobs (#4025)
## Summary

Will reduce the number of CI jobs we're running so that we:

- run all jobs on release PRs
- have some trigger to manually run all non-essential jobs based on
something (could be a comment, a tag, a webhook)
- then on regular commits run only essential jobs

## Details: 
- Creates a new "run-manual-tests" workflow, that can be triggered
through circleCI UI. I'll add a separate github action as a follow-up PR
so that we can also trigger it via a comment in a PR.
- Specifies the Xcode version explicitly for jobs in the job definition.
I saw that we were always specifying it when calling the job, which
would lead to some duplication, and in some cases it made the job by
itself not make a lot of sense - if a job is called "run tests in xcode
14", the xcode version shouldn't really be a param
- Adds a new job `pod-lib-lint`, which calls Cocoapods's `pod lib lint`.
This is a very helpful job in that it runs checks on all deployment
targets / OS versions / platforms, so it can catch a lot of wrong
availability checks that we might miss. It was already a part of some of
the other workflows, so this just makes it more explicit and removes it
from the places where it was left redundant.
- Also adds installation tests as part of the checks we run so that we
catch some other error early. These were previously only used in the
deploy workflow.