added new workflow to trigger all tests (#4051)
Follow up to #4025 

Adds a github action to trigger all tests based on a github PR comment. 

Annoyingly, this can only be [tested once the action is in the default
branch](https://docs.github.com/en/actions/using-workflows/events-that-trigger-workflows#issue_comment-on-issues-only-or-pull-requests-only),
the docs state:

> Note: This event will only trigger a workflow run if the workflow file
is on the default branch.

(I commented on this PR like 10 times before seeing this and didn't
understand why it wouldn't fire)

This uses [CircleCI's Marketplace action for
it](https://github.com/marketplace/actions/trigger-circleci-pipeline).

I also considered [triggering via
API](https://circleci.com/docs/triggers-overview/#run-a-pipeline-using-the-api),
but that has one big disadvantage in that the branch that the action
receives is `main`, so we'd have to do extra work to get the branch
name. This takes care of that automatically.