Customer Center compilation flag (#4149)
We have a long-running integration branch for Customer Center, long
running integration branches usually get very tricky to merge as time
passes on.

This PR attempts to solve it by creating a single custom compiler flag
that enables the Customer Center feature.

Here is what the compilation flag looks like: 
<img width="1107" alt="image"
src="https://github.com/user-attachments/assets/67df3bde-4c99-44b7-b33a-5847863a028c">


Then all we need to do to integrate into main is to delete that flag.
And re-add it for development. Easy peasy.

And when we want to ship Customer Center for real, we just delete all
usages of that flag.

If the flag isn't defined, then the code evaluates to false and is
simply skipped by the compiler. If it evaluates to true, then the code
runs.

---------

Co-authored-by: Toni Rico <antonio.rico.diez@revenuecat.com>